fix test

+++ app/controllers/likes_controller.rb
@@ -16,6 +16,9 @@ class LikesController < ApplicationController
 
   def destroy
     like = PostLike.find_by(post_id: @post.id, user_id: current_user.id)
+
+    redirect_to new_user_session_url if like.nil?
+
     if like.destroy
       redirect_to @post
     else
@@ -30,6 +33,6 @@ class LikesController < ApplicationController
   end
 
   def set_post
-    @post = Post.find(params[:post_id])
+    @post = Post.find_by(params[:post_id])
   end
 end
